Move frame-registration into content-all and encapsulate setup
src/dom/frame-registration.ts depended on chrome.* and was only used
from content-all, so it did not match the pure-DOM layer. Move it to
src/content-all/ and introduce setupFrameRegistration() that owns the
Maps, the onChildFrameId listener, the MutationObserver, the parent
announcement, and the frame-lifetime port.

Guard the chrome.runtime.connect call with `parent !== window`: the
root frame never sends RegisterChildFrame and tab close is already
handled by chrome.tabs.onRemoved, so the port serves no purpose there.
