+++ package.json
@@ -16,7 +16,7 @@
     "fix": "run-s fix:eslint fix:prettier",
     "fix:eslint": "eslint --flag unstable_native_nodejs_ts_config --cache --cache-strategy content --fix .",
     "fix:prettier": "prettier --write --log-level error --cache --cache-strategy content .",
-    "test": "tsx --test test/lib/*.test.ts test/background/*.test.ts test/dom/*.test.ts",
+    "test": "tsx --test test/lib/*.test.ts test/background/*.test.ts test/dom/*.test.ts test/content-all/*.test.ts",
     "test:e2e": "playwright test",
     "check": "run-s lint test",
     "watch": "run-p watch:build watch:serve",
