+++ tsconfig.json
@@ -11,6 +11,7 @@
     { "path": "test/lib/tsconfig.json" },
     { "path": "test/dom/tsconfig.json" },
     { "path": "test/background/tsconfig.json" },
+    { "path": "test/content-all/tsconfig.json" },
     { "path": "test/e2e/tsconfig.json" },
     { "path": "scripts/tsconfig.json" }
   ]
